frog now moves directly with lilypad, tracking if frog hops to one or zero. TODO: HUD, extra animation, win condition, and add other game modes/achievements

+++ Binary Bog/src/com/zxtodaro/binarybog/Frog.java
@@ -6,8 +6,8 @@ import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 
 public class Frog {
-	private int X;
-	private int Y;
+	private float X;
+	private float Y;
 	private int BoundX;
 	private int BoundY;
 	private Bitmap bmp;
@@ -38,13 +38,13 @@ public class Frog {
     		Y += speed * (runTime * 5f);
     }
 	
-	public int getX() {
+	public float getX() {
 		return X;
 	}
 	public void setX(int x) {
 		this.X = x;
 	}
-	public int getY() {
+	public float getY() {
 		return Y;
 	}
 	public void setY(int y) {
